backend/requirements.txt
# Flask Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Authentication
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.1.2

# Payment Processing
razorpay==1.4.2

# Environment & Config
python-dotenv==1.0.0

# Validation
pydantic==2.5.3
email-validator==2.1.0

# Utilities
python-multipart==0.0.6

# Production Server
gunicorn==21.2.0

# Image Upload & CDN
cloudinary==1.41.0
Pillow==10.1.0

# Optional: Caching & Task Queue
redis==5.0.1
celery==5.3.4
